+++ tests/Util/VariantBuilder.php
@@ -4,6 +4,7 @@ namespace PhpBench\Tests\Util;
 
 use DateTime;
 use PhpBench\Model\Benchmark;
+use PhpBench\Model\Error;
 use PhpBench\Model\ParameterSet;
 use PhpBench\Model\Subject;
 use PhpBench\Model\Suite;
@@ -31,6 +32,11 @@ final class VariantBuilder
      * @var string
      */
     private $name;
+
+    /**
+     * @var Error[]
+     */
+    private $errors;
 
     public function __construct(?SubjectBuilder $subjectBuilder, string $name)
     {
@@ -75,6 +81,12 @@ final class VariantBuilder
             $iteration->build($variant);
         }
 
+        $variant->computeStats();
+
+        if ($this->errors) {
+            $variant->createErrorStack($this->errors);
+        }
+
         return $variant;
     }
 
@@ -92,5 +104,12 @@ final class VariantBuilder
     public static function forSubjectBuilder(SubjectBuilder $subjectBuilder, string $name): self
     {
         return new self($subjectBuilder, $name);
+    }
+
+    public function withError(Error $error): self
+    {
+        $this->errors[] = $error;
+
+        return $this;
     }
 }